Revert changes to spill-over examples.
I didn't make those changes in the first place on a whim.

The left-alignment was applied because it is consistent with the direction of the paragraph, and is a cue to the reader that the context here is LTR.

The و" CSS" was removed from the example because (a) it has the possibility of introducing additional wierdness that is not related to the point of this example, (b) because it makes the example long and more likely to wrap, which makes it harder to grok the example and in some cases could introduces other unrelated issues for the general reader, (d) it makes it easier to edit the example, and (d) because it doesn't matter that it's not exactly the same as examples used elsewhere – it's just an example.