Reorganise the Best Practices section
I've been struggling with this section for a long time.  I finally sat down and thought through how we could make it read better, and reduce duplication.  This is my proposal.

I mostly just moved things around.  There is only a very small amount of text change.  It will probably be necessary to go through the text one more time to ensure that the narrative flows well.

I think this makes it easier to grock the key recommendations, and reduces previous duplications which were a little confusing.