Rewrite of initial recommendations in guidelines
We need to start out by saying that we propose use of metadata, and whether or not a resource-wide option is available, there is always a need for string-specific metadata.  So it makes sense to move that requirement to the beginning of the section.  Then propose an optional resource-wide metadata declaration, in addition.  Then require that in the absence of metadata, language & direction are unknown, so that consumers can apply appropriate heuristics.